requirements.txt
djangorestframework
Pillow